+++ changelog.md
@@ -9,8 +9,7 @@
 + Added sku fallback for products without names.
 + Updated translations.
 + Added voucher payment method.
-+ Bugfix - phone numbers for billing and delivery address are sent with international format and regular phone numbers are tried to be converted to it. If convertion
-  fails in any way then phone number is not sent and thus an informative message is logged in PrestaShop logger.
++ Phone numbers without possible international code ( which does not have + in front of the number ) are not sent to mollie as delivery or billing address phone.
 
 ## Changes in release 4.0.9 Hotfix-1 ##
 + In PS1.6 prevents double click payment method to create multiple orders